requirements.txt
pyyaml
bitly_api
pyinotify

